+++ CHANGELOG.md
@@ -1,3 +1,46 @@
+## [71.1.2](https://github.com/elastic/elastic-charts/compare/v71.1.1...v71.1.2) (2025-11-07)
+
+
+### Bug Fixes
+
+* **bullet:** fix metric breakpoints in bullet chart ([#2759](https://github.com/elastic/elastic-charts/issues/2759)) ([84ad335](https://github.com/elastic/elastic-charts/commit/84ad335316a59955f300d912155421058162ff27))
+* **highlighter:** Fix sorting logic for highlighted geometries ([#2756](https://github.com/elastic/elastic-charts/issues/2756)) ([6054ca1](https://github.com/elastic/elastic-charts/commit/6054ca176cc747c980ce398b99630886188ea942))
+
+
+### Performance Improvements
+
+* **xy:** avoid computing bar value display width ([#2754](https://github.com/elastic/elastic-charts/issues/2754)) ([26b1d65](https://github.com/elastic/elastic-charts/commit/26b1d657cb3db691105b722fe081f0c6145522e0))
+
+## [71.1.1](https://github.com/elastic/elastic-charts/compare/v71.1.0...v71.1.1) (2025-09-29)
+
+
+### Bug Fixes
+
+* remove aria-label form div (part 2) ([#2751](https://github.com/elastic/elastic-charts/issues/2751)) ([602c743](https://github.com/elastic/elastic-charts/commit/602c743def7ebc551ce2a56ad6d54d42647c8506))
+* remove aria-label from div ([#2750](https://github.com/elastic/elastic-charts/issues/2750)) ([e8e22d8](https://github.com/elastic/elastic-charts/commit/e8e22d8a2f4e1d7f85b9c1543ba6324bdeb7582a))
+
+# [71.1.0](https://github.com/elastic/elastic-charts/compare/v71.0.1...v71.1.0) (2025-09-25)
+
+
+### Bug Fixes
+
+* **metric:** remove support for small responsive breakpoints ([#2743](https://github.com/elastic/elastic-charts/issues/2743)) ([a5132d9](https://github.com/elastic/elastic-charts/commit/a5132d9292f1e135c948dd6adfa1f14a79232d59))
+* remove wrong traversal path in imports ([#2747](https://github.com/elastic/elastic-charts/issues/2747)) ([0f55808](https://github.com/elastic/elastic-charts/commit/0f55808192afe2caef88e03f19841ef2a8a31e17))
+
+
+### Features
+
+* **metric:** add ability to disable secondary metric badge border ([#2734](https://github.com/elastic/elastic-charts/issues/2734)) ([cb04bf5](https://github.com/elastic/elastic-charts/commit/cb04bf5d44d952120e25a3e9528e508c60ec5a56))
+
+## [71.0.1](https://github.com/elastic/elastic-charts/compare/v71.0.0...v71.0.1) (2025-09-22)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @playwright/test to ^1.54.2 ([#2553](https://github.com/elastic/elastic-charts/issues/2553)) ([a75c59c](https://github.com/elastic/elastic-charts/commit/a75c59cdcad2bdf3a3ef43c20132c45085412ed7))
+* **metric:** fix `fit` font size mode ([#2730](https://github.com/elastic/elastic-charts/issues/2730)) ([98b4204](https://github.com/elastic/elastic-charts/commit/98b420485fdb09d64209e3056d206b3fc6733d55))
+* **partition:** fix tooltip nested colors  ([#2717](https://github.com/elastic/elastic-charts/issues/2717)) ([591459e](https://github.com/elastic/elastic-charts/commit/591459ed8e8aa56a4a03914427a65b0e400d7a8a))
+
 # [71.0.0](https://github.com/elastic/elastic-charts/compare/v70.1.3...v71.0.0) (2025-08-11)